third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-feature-settings-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-feature-settings value '"tnum", "hist"'
  assert_equals: expected "\\"hist\\", \\"tnum\\"" but got "\\"tnum\\", \\"hist\\""
[FAIL] Property font-feature-settings value '"dlig", "smcp", "dlig" 0'
  assert_equals: expected "\\"dlig\\" 0, \\"smcp\\"" but got "\\"dlig\\", \\"smcp\\", \\"dlig\\" 0"
[FAIL] Property font-feature-settings value '"liga" calc(10 + (sign(2cqw - 10px) * 5)), "dlig" calc(20 + (sign(2cqw - 10px) * 5))'
  assert_equals: expected "\\"dlig\\" 15, \\"liga\\" 5" but got "\\"liga\\" 5, \\"dlig\\" 15"
Harness: the test ran to completion.
